+++ frontend/src/app/pages/policy/legal.pages.ts
@@ -33,7 +33,7 @@ export const legalSlots: LegalPagesConfig = {
  */
 export const legalUpdated: Partial<Record<LegalPageId, Date>> = {
     [LegalPages.PrivacyPolicy]: new Date('2026-07-03'),
-    [LegalPages.CookiePolicy]: new Date('2026-08-18'),
+    [LegalPages.CookiePolicy]: new Date('2026-08-20'),
     [LegalPages.TermsOfService]: new Date('2026-07-03'),
     [LegalPages.LegalNotice]: new Date('2026-07-03'),
     [LegalPages.AccessibilityStatement]: new Date('2026-07-08'),